backend/requirements.txt
python-dotenv
sqlalchemy
psycopg2-binary
fastapi
uvicorn